+++ libhoomd/updaters/TwoStepBDNVTRigid.cc
@@ -73,13 +73,17 @@ TwoStepBDNVTRigid::TwoStepBDNVTRigid(boost::shared_ptr<SystemDefinition> sysdef,
                                      boost::shared_ptr<ParticleGroup> group,
                                      boost::shared_ptr<Variant> T,
                                      unsigned int seed,
-                                     bool gamma_diam)
-    : TwoStepNVERigid(sysdef, group), m_seed(seed), m_gamma_diam(gamma_diam)
+                                     bool gamma_diam,
+                                     bool noiseless_t,
+                                     bool noiseless_r
+                                    )
+    : TwoStepNVERigid(sysdef, group), m_seed(seed), m_gamma_diam(gamma_diam), m_gamma_r(0),
+      m_noiseless_r(noiseless_r), m_noiseless_t(noiseless_t)
     {
     m_exec_conf->msg->notice(5) << "Constructing TwoStepBDNVTRigid" << endl;
 
     m_temperature = T;
- 
+
     // set a named, but otherwise blank set of integrator variables
     IntegratorVariables v = getIntegratorVariables();
 
@@ -149,6 +153,113 @@ void TwoStepBDNVTRigid::setGamma(unsigned int typ, Scalar gamma)
     h_gamma.data[typ] = gamma;
     }
 
+
+/*! \param typ Particle type to set gamma for
+    \param gamma The gamma value to set
+*/
+void TwoStepBDNVTRigid::setGamma_r(Scalar gamma_r)
+    {
+    // check for user errors
+    m_gamma_r = gamma_r;
+    }
+
+
+/* override NVERigid Step One
+*/
+void TwoStepBDNVTRigid::integrateStepOne(unsigned int timestep)
+    {
+    if (m_first_step)
+        {
+        setup();
+        m_first_step = false;
+        }
+
+    // sanity check
+    if (m_n_bodies <= 0)
+        return;
+
+    if (m_prof)
+        m_prof->push("NVE rigid step 1");
+
+    // get box
+    const BoxDim& box = m_pdata->getBox();
+    const Scalar D = Scalar(m_sysdef->getNDimensions());
+
+    // now we can get on with the velocity verlet: initial integration
+    {
+    // rigid data handles
+    ArrayHandle<Scalar> body_mass_handle(m_rigid_data->getBodyMass(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> moment_inertia_handle(m_rigid_data->getMomentInertia(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> force_handle(m_rigid_data->getForce(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> torque_handle(m_rigid_data->getTorque(), access_location::host, access_mode::read);
+
+    ArrayHandle<Scalar4> com_handle(m_rigid_data->getCOM(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> vel_handle(m_rigid_data->getVel(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> orientation_handle(m_rigid_data->getOrientation(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> angmom_handle(m_rigid_data->getAngMom(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> angvel_handle(m_rigid_data->getAngVel(), access_location::host, access_mode::readwrite);
+
+    ArrayHandle<int3> body_image_handle(m_rigid_data->getBodyImage(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> ex_space_handle(m_rigid_data->getExSpace(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> ey_space_handle(m_rigid_data->getEySpace(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar4> ez_space_handle(m_rigid_data->getEzSpace(), access_location::host, access_mode::readwrite);
+
+    // Scalar h_gamma_r (m_gamma_r);
+
+    Scalar dt_half = 0.5 * m_deltaT;
+    Scalar dtfm;
+
+    // initialize the RNG
+    Saru saru(m_seed, timestep);
+
+    // for each body
+    for (unsigned int group_idx = 0; group_idx < m_n_bodies; group_idx++)
+        {
+        unsigned int body = m_body_group->getMemberIndex(group_idx);
+
+        dtfm = dt_half / body_mass_handle.data[body];
+        vel_handle.data[body].x += dtfm * force_handle.data[body].x;
+        vel_handle.data[body].y += dtfm * force_handle.data[body].y;
+        vel_handle.data[body].z += dtfm * force_handle.data[body].z;
+
+        com_handle.data[body].x += vel_handle.data[body].x * m_deltaT;
+        com_handle.data[body].y += vel_handle.data[body].y * m_deltaT;
+        com_handle.data[body].z += vel_handle.data[body].z * m_deltaT;
+
+        box.wrap(com_handle.data[body], body_image_handle.data[body]);
+
+        // // added:
+        // if (D < 3.0)
+        //     {
+        //     torque_handle.data[body].x -= h_gamma_r * angvel_handle.data[body].x;
+        //     torque_handle.data[body].y -= h_gamma_r * angvel_handle.data[body].y;
+        //     torque_handle.data[body].z -= h_gamma_r * angvel_handle.data[body].z;
+        //     }
+
+        // update the angular momentum
+        angmom_handle.data[body].x += dt_half * torque_handle.data[body].x;
+        angmom_handle.data[body].y += dt_half * torque_handle.data[body].y;
+        angmom_handle.data[body].z += dt_half * torque_handle.data[body].z;
+
+        // update quaternion and angular velocity
+        advanceQuaternion(angmom_handle.data[body],
+                          moment_inertia_handle.data[body],
+                          angvel_handle.data[body],
+                          ex_space_handle.data[body],
+                          ey_space_handle.data[body],
+                          ez_space_handle.data[body],
+                          m_deltaT,
+                          orientation_handle.data[body]);
+        }
+    } // out of scope for handles
+
+
+    if (m_prof)
+        m_prof->pop();
+    }
+
+
+
 /*! \param timestep Current time step
     \post particle velocities are moved forward to timestep+1
 */
@@ -164,64 +275,71 @@ void TwoStepBDNVTRigid::integrateStepTwo(unsigned int timestep)
 
     const GPUArray< Scalar4 >& net_force = m_pdata->getNetForce();
 
-    {
-    // Modify the net forces with the random and drag forces
-    ArrayHandle<Scalar4> h_vel(m_pdata->getVelocities(), access_location::host, access_mode::readwrite);
-    ArrayHandle<Scalar3> h_accel(m_pdata->getAccelerations(), access_location::host, access_mode::readwrite);
 
-    ArrayHandle<Scalar4> h_net_force(net_force, access_location::host, access_mode::readwrite);
-    ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::read);
+    // skip all random and damping force update before the rigid body force and torque calculation
+    // if (0)
+    //     {
+    //     // Modify the net forces with the random and drag forces
+    //     ArrayHandle<Scalar4> h_vel(m_pdata->getVelocities(), access_location::host, access_mode::readwrite);
+    //     ArrayHandle<Scalar3> h_accel(m_pdata->getAccelerations(), access_location::host, access_mode::readwrite);
 
-    ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(), access_location::host, access_mode::read);
-    ArrayHandle<Scalar4> h_pos(m_pdata->getPositions(), access_location::host, access_mode::read);
+    //     ArrayHandle<Scalar4> h_net_force(net_force, access_location::host, access_mode::readwrite);
+    //     ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::read);
 
-    // grab some initial variables
-    const Scalar currentTemp = m_temperature->getValue(timestep);
-    const Scalar D = Scalar(m_sysdef->getNDimensions());
+    //     ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(), access_location::host, access_mode::read);
+    //     ArrayHandle<Scalar4> h_pos(m_pdata->getPositions(), access_location::host, access_mode::read);
 
-    // initialize the RNG
-    Saru saru(m_seed, timestep);
+    //     // grab some initial variables
+    //     const Scalar currentTemp = m_temperature->getValue(timestep);
+    //     const Scalar D = Scalar(m_sysdef->getNDimensions());
 
-    // a(t+deltaT) gets modified with the bd forces
-    // v(t+deltaT) = v(t+deltaT/2) + 1/2 * a(t+deltaT)*deltaT
-    unsigned int group_size = m_group->getNumMembers();
-    for (unsigned int group_idx = 0; group_idx < group_size; group_idx++)
-        {
-        unsigned int j = m_group->getMemberIndex(group_idx);
+    //     // initialize the RNG
+    //     Saru saru(m_seed, timestep);
 
-        // first, calculate the BD forces
-        // Generate three random numbers
-        Scalar rx = saru.d(-1,1);
-        Scalar ry = saru.d(-1,1);
-        Scalar rz = saru.d(-1,1);
+    //     // a(t+deltaT) gets modified with the bd forces
+    //     // v(t+deltaT) = v(t+deltaT/2) + 1/2 * a(t+deltaT)*deltaT
 
-        Scalar gamma;
-        if (m_gamma_diam)
-            gamma = h_diameter.data[j];
-        else
-            {
-            unsigned int type = __scalar_as_int(h_pos.data[j].w);
-            gamma = h_gamma.data[type];
-            }
+    //     unsigned int group_size = m_group->getNumMembers();
+    //     for (unsigned int group_idx = 0; group_idx < group_size; group_idx++)
+    //         {
+    //         unsigned int j = m_group->getMemberIndex(group_idx);
 
-        // compute the bd force
-        Scalar coeff = sqrt(Scalar(6.0)*gamma*currentTemp/m_deltaT);
-        Scalar bd_fx = rx*coeff - gamma*h_vel.data[j].x;
-        Scalar bd_fy = ry*coeff - gamma*h_vel.data[j].y;
-        Scalar bd_fz = rz*coeff - gamma*h_vel.data[j].z;
+    //         // first, calculate the BD forces
+    //         // Generate three random numbers
+    //         Scalar rx = saru.d(-1,1);
+    //         Scalar ry = saru.d(-1,1);
+    //         Scalar rz = saru.d(-1,1);
 
-        if (D < 3.0)
-            bd_fz = Scalar(0.0);
+    //         Scalar gamma;
+    //         if (m_gamma_diam)
+    //             gamma = h_diameter.data[j];
+    //         else
+    //             {
+    //             unsigned int type = __scalar_as_int(h_pos.data[j].w);
+    //             gamma = h_gamma.data[type];
+    //             }
 
-        h_net_force.data[j].x += bd_fx;
-        h_net_force.data[j].y += bd_fy;
-        h_net_force.data[j].z += bd_fz;
-        }
-    }
+    //         // compute the bd force
+    //         Scalar coeff = sqrt(Scalar(6.0)*gamma*currentTemp/m_deltaT);
+    //         Scalar bd_fx = rx*coeff - gamma*h_vel.data[j].x;
+    //         Scalar bd_fy = ry*coeff - gamma*h_vel.data[j].y;
+    //         Scalar bd_fz = rz*coeff - gamma*h_vel.data[j].z;
+
+    //         if (D < 3.0)
+    //             bd_fz = Scalar(0.0);
+
+    //         h_net_force.data[j].x += bd_fx;
+    //         h_net_force.data[j].y += bd_fy;
+    //         h_net_force.data[j].z += bd_fz;
+    //         }
+    //     }
 
     // Perform the second step like in TwoStepNVERigid
     // compute net forces and torques on rigid bodies from particle forces
     computeForceAndTorque(timestep);
+
+
+
 
     {
     // rigid data handes
@@ -259,6 +377,92 @@ void TwoStepBDNVTRigid::integrateStepTwo(unsigned int timestep)
         }
     } // out of scope for handles
 
+
+
+
+    // added: resume the random and damping force calculation
+    {
+    // Modify the net forces with the random and drag forces
+    ArrayHandle<Scalar4> h_vel(m_pdata->getVelocities(), access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar3> h_accel(m_pdata->getAccelerations(), access_location::host, access_mode::readwrite);
+
+    ArrayHandle<Scalar4> h_net_force(net_force, access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::read);
+
+    ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> h_pos(m_pdata->getPositions(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> torque_handle(m_rigid_data->getTorque(), access_location::host, access_mode::read);
+    ArrayHandle<Scalar4> angvel_handle(m_rigid_data->getAngVel(), access_location::host, access_mode::readwrite);
+
+    // grab some initial variables
+    const Scalar currentTemp = m_temperature->getValue(timestep);
+    const Scalar D = Scalar(m_sysdef->getNDimensions());
+
+    // initialize the RNG
+    Saru saru(m_seed, timestep);
+
+    // for each particle
+    unsigned int group_size = m_group->getNumMembers();
+    for (unsigned int group_idx = 0; group_idx < group_size; group_idx++)
+        {
+        unsigned int j = m_group->getMemberIndex(group_idx);
+
+        // first, calculate the BD forces
+        // Generate three random numbers
+        Scalar rx = saru.d(-1,1);
+        Scalar ry = saru.d(-1,1);
+        Scalar rz = saru.d(-1,1);
+
+        Scalar gamma;
+        if (m_gamma_diam)
+            gamma = h_diameter.data[j];
+        else
+            {
+            unsigned int type = __scalar_as_int(h_pos.data[j].w);
+            gamma = h_gamma.data[type];
+            }
+
+        // compute the bd force
+        Scalar coeff = sqrt(Scalar(6.0)*gamma*currentTemp/m_deltaT);
+        if (m_noiseless_t)  coeff = Scalar(0.0);
+
+        Scalar bd_fx = rx*coeff - gamma*h_vel.data[j].x;
+        Scalar bd_fy = ry*coeff - gamma*h_vel.data[j].y;
+        Scalar bd_fz = rz*coeff - gamma*h_vel.data[j].z;
+
+        if (D < 3.0)
+            bd_fz = Scalar(0.0);
+
+        h_net_force.data[j].x += bd_fx;
+        h_net_force.data[j].y += bd_fy;
+        h_net_force.data[j].z += bd_fz;
+        }
+
+    // for each body
+    Scalar h_gamma_r (m_gamma_r);
+    for (unsigned int group_idx = 0; group_idx < m_n_bodies; group_idx++)
+        {
+        unsigned int body = m_body_group->getMemberIndex(group_idx);
+
+        if (D < 3.0)
+            {
+            Scalar rrx = saru.d(-1,1);
+            Scalar rry = saru.d(-1,1);
+            Scalar rrz = saru.d(-1,1);
+            Scalar coeff_r = sqrt(Scalar(6.0)*h_gamma_r*currentTemp/m_deltaT);
+            if (m_noiseless_r)  coeff_r = Scalar(0.0);
+
+            torque_handle.data[body].x -= h_gamma_r * angvel_handle.data[body].x;
+            torque_handle.data[body].y -= h_gamma_r * angvel_handle.data[body].y;
+            // torque_handle.data[body].x += coeff_r * rrx - h_gamma_r * angvel_handle.data[body].x;
+            // torque_handle.data[body].y += coeff_r * rry - h_gamma_r * angvel_handle.data[body].y;
+            torque_handle.data[body].z += coeff_r * rrz - h_gamma_r * angvel_handle.data[body].z;
+            }
+        }
+    }
+
+
+
     // done profiling
     if (m_prof)
         m_prof->pop();
@@ -271,8 +475,11 @@ void export_TwoStepBDNVTRigid()
                          boost::shared_ptr<ParticleGroup>,
                          boost::shared_ptr<Variant>,
                          unsigned int,
+                         bool,
+                         bool,
                          bool
                          >())
         .def("setGamma", &TwoStepBDNVTRigid::setGamma)
+        .def("setGamma_r", &TwoStepBDNVTRigid::setGamma_r)
         ;
     }