+++ ChangeLog.md
@@ -7,6 +7,9 @@
 *New features*
 
 * Distance constraints `constrain.distance` - constrain pairs of particles to a fixed separation distance
+* `force.active` applies an active force (optionally with rotational diffusion) to a group of particles
+* `update.constrain_ellipsoid` constrains particles to an ellipsoid
+* `integrate.langevin` and `ingetgrate.brownian` now apply rotational noise and damping to anisotropic particles
 
 *Bug fixes*